# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: MSHOPPA API
  version: 0.1.0
paths:
...
  /api/checkout/:
    post:
      operationId: checkout_create
      tags:
      - checkout
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
        required: true
      security:
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
          description: ''
...
  /api/payments/checkout/:
    post:
      operationId: payments_checkout_create
      tags:
      - payments
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HandoffInputRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/HandoffInputRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/HandoffInputRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: ''
        '201':
          content:
            application/json:
              schema:
                type: object
          description: ''
...
components:
  schemas:
...
    CartItemRequest:
      type: object
      properties:
        variant_id:
          type: string
          format: uuid
        quantity:
          type: integer
          maximum: 99
          minimum: 1
      required:
      - quantity
      - variant_id
    CheckoutRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItemRequest'
        checkout_key:
          type: string
          format: uuid
        quote_token:
          type: string
          minLength: 1
          maxLength: 2000
        name:
          type: string
          minLength: 1
          maxLength: 150
        email:
          type: string
          format: email
          minLength: 1
        phone:
          type: string
          default: ''
          maxLength: 30
        address:
          type: string
          minLength: 1
          maxLength: 500
        provider:
          $ref: '#/components/schemas/ProviderEnum'
      required:
      - address
      - checkout_key
      - email
      - items
      - name
      - provider
      - quote_token
...
    HandoffInputRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
          maxLength: 16000
        details:
          type: object
          additionalProperties: {}
      required:
      - token
...
    ProviderEnum:
      enum:
      - stripe
      - mpesa
      type: string
      description: |-
        * `stripe` - stripe
        * `mpesa` - mpesa
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mshoppa_session